backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
PyPDF2==3.0.1
pdfplumber==0.10.0
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
tiktoken==0.5.1
